# distsql: add external storage to routers