Closes #25764

Still need to add tests and benchmark (iterator creation might not be cheap) but wanted to get some feedback.