@jordanlewis 2018-06-11
Reviewed 2 of 2 files at r1, 1 of 1 files at r2, 1 of 1 files at r3.
Review status: :shipit: complete! 0 of 0 LGTMs obtained

---

*[pkg/sql/distsqlrun/disk_row_container.go, line 266 at r2](https://reviewable.io/reviews/cockroachdb/cockroach/26604#-LEkRRMU93AWAjmPe7cF:-LEkRRMU93AWAjmPe7cG:b-huhjbj) ([raw file](https://github.com/cockroachdb/cockroach/blob/5c0ce8f71edb14eb7d5c376891aae5d081eca82c/pkg/sql/distsqlrun/disk_row_container.go#L266)):*
> ```Go
> 		return nil, err
> 	}
> 	r.diskRowIterator.rowContainer.lastReadKey = r.Key()
> ```

I think you can `UnsafeKey` here, since we call this on every `Row()` invocation.

---

*[pkg/sql/distsqlrun/routers.go, line 133 at r3](https://reviewable.io/reviews/cockroachdb/cockroach/26604#-LEkRlSz3rkgOdy1PfZ-:-LEkRlSz3rkgOdy1PfZ0:b-4hzxt1) ([raw file](https://github.com/cockroachdb/cockroach/blob/e598d1f59345c67024145c2c17691907c0041160/pkg/sql/distsqlrun/routers.go#L133)):*
> ```Go
> 	// First try to get rows from the row container.
> 	if ro.mu.rowContainer.Len() > 0 {
> 		if err := func() error {
> ```

I think this might cause allocations... can you make sure it doesn't?

---

*[pkg/sql/distsqlrun/routers.go, line 136 at r3](https://reviewable.io/reviews/cockroachdb/cockroach/26604#-LEkRvngBuGogJNCULIg:-LEkRvngBuGogJNCULIh:b-7l71p4) ([raw file](https://github.com/cockroachdb/cockroach/blob/e598d1f59345c67024145c2c17691907c0041160/pkg/sql/distsqlrun/routers.go#L136)):*
> ```Go
> 			i := ro.mu.rowContainer.NewFinalIterator(ctx)
> 			defer i.Close()
> 			for i.Rewind(); n < len(rowBuf); i.Next() {
> ```

Why does this need to be rewound? Aren't we only using it here, in which case we should be able to just call Next?

---

*[pkg/sql/distsqlrun/routers.go, line 156 at r3](https://reviewable.io/reviews/cockroachdb/cockroach/26604#-LEkS7JO-Ec1-6SFiLa9:-LEkS7JO-Ec1-6SFiLaA:b3kre) ([raw file](https://github.com/cockroachdb/cockroach/blob/e598d1f59345c67024145c2c17691907c0041160/pkg/sql/distsqlrun/routers.go#L156)):*
> ```Go
> 		}
> 	}
> 	for ; n < len(rowBuf) && ro.mu.rowContainer.Len() > 0; n++ {
> ```

?

---


*Comments from [Reviewable](https://reviewable.io/reviews/cockroachdb/cockroach/26604)*
<!-- Sent from Reviewable.io -->