@asubiotto 2018-06-14
Getting everything from
```
name                        old time/op    new time/op    delta
Router/BY_HASH/outputs=2-8    18.3ms ± 1%    20.7ms ± 1%  +13.30%  (p=0.000 n=10+9)
Router/BY_HASH/outputs=4-8    15.1ms ± 1%    16.9ms ± 2%  +12.39%  (p=0.000 n=10+10)
Router/BY_HASH/outputs=8-8    22.2ms ± 2%    24.1ms ± 4%   +8.76%  (p=0.000 n=8+10)

name                        old speed      new speed      delta
Router/BY_HASH/outputs=2-8  57.4MB/s ± 1%  50.7MB/s ± 1%  -11.74%  (p=0.000 n=10+9)
Router/BY_HASH/outputs=4-8   139MB/s ± 1%   124MB/s ± 2%  -11.02%  (p=0.000 n=10+10)
Router/BY_HASH/outputs=8-8   189MB/s ± 2%   174MB/s ± 4%   -8.01%  (p=0.000 n=8+10)

name                        old alloc/op   new alloc/op   delta
Router/BY_HASH/outputs=2-8     680kB ± 2%     500kB ± 3%  -26.36%  (p=0.000 n=10+9)
Router/BY_HASH/outputs=4-8     203kB ± 2%     128kB ± 1%  -36.94%  (p=0.000 n=9+9)
Router/BY_HASH/outputs=8-8    44.0kB ± 3%  176.6kB ±329%     ~     (p=0.400 n=9+10)

name                        old allocs/op  new allocs/op  delta
Router/BY_HASH/outputs=2-8     10.6k ± 2%      7.8k ± 3%  -26.86%  (p=0.000 n=10+9)
Router/BY_HASH/outputs=4-8     2.95k ± 4%     1.75k ± 1%  -40.86%  (p=0.000 n=10+9)
Router/BY_HASH/outputs=8-8       228 ±10%     2300 ±392%     ~     (p=0.345 n=9+10)
```
There's a performance hit but the only thing that we're doing differently (because we're only benchmarking the in-memory case) is creating this final iterator struct and doing some extra hops through a `diskBackedRowContainer`. I think that this is inevitable if we want to keep the code simple and have external storage behavior.

I spent a while going down the `Key/UnsafeKey` rabbit hole but at least it produced some good tests. Replied with more detail to your comment.

---

Review status: :shipit: complete! 0 of 0 LGTMs obtained

---

*[pkg/sql/distsqlrun/disk_row_container.go, line 266 at r2](https://reviewable.io/reviews/cockroachdb/cockroach/26604#-LEkRRMU93AWAjmPe7cF:-LEpWb3nBtv5de1JTYPS:b-pnhf5t) ([raw file](https://github.com/cockroachdb/cockroach/blob/5c0ce8f71edb14eb7d5c376891aae5d081eca82c/pkg/sql/distsqlrun/disk_row_container.go#L266)):*
<details><summary><i>Previously, jordanlewis (Jordan Lewis) wrote…</i></summary><blockquote>

I think you can `UnsafeKey` here, since we call this on every `Row()` invocation.
</blockquote></details>

I tried removing the allocation for the `Key` here but I think it's too complicated for now. The issue is that if we use`UnsafeKey` here, that memory will be changed from under us by any call to `Next`. That made me think it might actually be better to just call `UnsafeKey` in `Next`, but you're not guaranteed to have a valid key (e.g. you've gone past the end of the iteration) so you would have to get the key before the underlying call to next but you would have to make a copy anyway so that the `Next` call doesn't mess with the key which would leave you in the same spot.

I also tried to update the key in `Close` but that ran into the issue again of needing a last read key because of the case in which you close and have reached the end, but might add more rows and want to resume from there. The only way to get that would to again make an allocation in `Next` to keep the last read key around. I think this is simpler but I added a TODO, let me know if you think I missed anything.

Another thing that this has uncovered is that we cannot use a disk-based final iterator with a diskRowContainer with an ordering unless we stop adding rows once we create a final iterator since We could add a row that sorts before the current row.

---

*[pkg/sql/distsqlrun/routers.go, line 133 at r3](https://reviewable.io/reviews/cockroachdb/cockroach/26604#-LEkRlSz3rkgOdy1PfZ-:-LEpbwW_2pHVhHSyfwRP:bx9ee55) ([raw file](https://github.com/cockroachdb/cockroach/blob/e598d1f59345c67024145c2c17691907c0041160/pkg/sql/distsqlrun/routers.go#L133)):*
<details><summary><i>Previously, jordanlewis (Jordan Lewis) wrote…</i></summary><blockquote>

I think this might cause allocations... can you make sure it doesn't?
</blockquote></details>

Used `scripts/goescape` to make sure it doesn't

---

*[pkg/sql/distsqlrun/routers.go, line 136 at r3](https://reviewable.io/reviews/cockroachdb/cockroach/26604#-LEkRvngBuGogJNCULIg:-LEymuZ7A9KuFz66k3ql:bydhi6u) ([raw file](https://github.com/cockroachdb/cockroach/blob/e598d1f59345c67024145c2c17691907c0041160/pkg/sql/distsqlrun/routers.go#L136)):*
<details><summary><i>Previously, jordanlewis (Jordan Lewis) wrote…</i></summary><blockquote>

Why does this need to be rewound? Aren't we only using it here, in which case we should be able to just call Next?
</blockquote></details>

Rewind` is convention when using iterators. In this case, the rewind is a seek to the key you actually want to start reading from. If you don't do it it'll probably start reading from the nil key although I think that reading without rewinding results in undefined behavior.

---

*[pkg/sql/distsqlrun/routers.go, line 156 at r3](https://reviewable.io/reviews/cockroachdb/cockroach/26604#-LEkS7JO-Ec1-6SFiLa9:-LEpdOBiAiGCuHWOHjkj:bb71nyg) ([raw file](https://github.com/cockroachdb/cockroach/blob/e598d1f59345c67024145c2c17691907c0041160/pkg/sql/distsqlrun/routers.go#L156)):*
<details><summary><i>Previously, jordanlewis (Jordan Lewis) wrote…</i></summary><blockquote>

?
</blockquote></details>

Oops

---


*Comments from [Reviewable](https://reviewable.io/reviews/cockroachdb/cockroach/26604#-:-LEpi2MpEglJiy3mV-Ju:b85g526)*
<!-- Sent from Reviewable.io -->